Issue #26: Moved parse_range from load_*.py to utils.py. The previous locations now import the function instead. Added final update for progress reports.
