+++ self_rec_framework/src/data_generation/data_loading/load_pku_saferlhf.py
@@ -41,7 +41,7 @@ import uuid
 from datasets import load_dataset
 
 from self_rec_framework.src.helpers.constants import MY_DATASET_NAMESPACE
-from self_rec_framework.src.helpers.utils import save_json, data_dir
+from self_rec_framework.src.helpers.utils import save_json, data_dir, parse_range
 
 
 def load_pku_saferlhf_data(
@@ -118,7 +118,7 @@ def load_pku_saferlhf_data(
         # Store in input dict
         input_dict[sample_uuid] = prompt
 
-        if (idx + 1) % 1000 == 0:
+        if (idx + 1) % 1000 == 0 or idx == len(dataset) - 1: # mod 1000 for progress updates, last sample for final update
             print(f"  Processed {idx + 1}/{len(dataset)} samples...")
 
     # Save to data/{dataset_name}/input.json
@@ -133,20 +133,6 @@ def load_pku_saferlhf_data(
     print(f"  - Samples: {len(input_dict)}")
     print("  - Content: Prompts/questions from SafeRLHF")
     print("  - Task: Question answering / dialogue")
-
-
-def parse_range(range_str: str) -> tuple[int, int]:
-    """Parse a range string like '5-15' into a tuple of (start, end)."""
-    try:
-        parts = range_str.split("-")
-        if len(parts) != 2:
-            raise ValueError
-        start, end = int(parts[0]), int(parts[1])
-        return (start, end)
-    except (ValueError, AttributeError):
-        raise argparse.ArgumentTypeError(
-            f"Range must be in format 'START-END' (e.g., '5-15'), got: {range_str}"
-        )
 
 
 def main():
