+++ self_rec_framework/src/data_generation/data_loading/load_sharegpt.py
@@ -32,7 +32,7 @@ import re
 from huggingface_hub import hf_hub_download
 
 from self_rec_framework.src.helpers.constants import MY_DATASET_NAMESPACE
-from self_rec_framework.src.helpers.utils import save_json, data_dir
+from self_rec_framework.src.helpers.utils import save_json, data_dir, parse_range
 
 
 def clean_html(text: str) -> str:
@@ -209,7 +209,7 @@ def load_sharegpt_data(
         else:
             skipped_count += 1
 
-        if (idx + 1) % 1000 == 0:
+        if (idx + 1) % 1000 == 0 or idx == len(all_samples) - 1: # mod 1000 for progress updates, last sample for final update
             print(f"  Processed {idx + 1}/{len(all_samples)} samples...")
 
     if skipped_count > 0:
@@ -225,20 +225,6 @@ def load_sharegpt_data(
     print(f"  - Samples: {len(input_dict)}")
     print("  - Content: First user prompts from ShareGPT conversations")
     print("  - Task: Dialogue / conversation")
-
-
-def parse_range(range_str: str) -> tuple[int, int]:
-    """Parse a range string like '5-15' into a tuple of (start, end)."""
-    try:
-        parts = range_str.split("-")
-        if len(parts) != 2:
-            raise ValueError
-        start, end = int(parts[0]), int(parts[1])
-        return (start, end)
-    except (ValueError, AttributeError):
-        raise argparse.ArgumentTypeError(
-            f"Range must be in format 'START-END' (e.g., '5-15'), got: {range_str}"
-        )
 
 
 def main():
